+++ app/src/app.rs
@@ -1,5 +1,6 @@
-use cult_common::*;
+use std::net::SocketAddr;
 use yew::prelude::*;
+use cult_common::parse_addr_str;
 
 use crate::websocket::WebsocketService;
