+++ cult-common/Cargo.toml
@@ -6,9 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-actix = "0.13.3"
-actix-web = "4.5.1"
-actix-web-actors = "4.3.0"
 anyhow = "1.0.82"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1"
