+++ server/src/gamewebsocket.rs
@@ -4,7 +4,7 @@ use actix::Addr;
 use actix_web::{HttpRequest, HttpResponse, web};
 use actix_web_actors::ws;
 use serde_json::json;
-use cult_common::{extract_value, get_internal_server_error_json};
+use crate::lib::{extract_value, get_internal_server_error_json};
 use crate::server;
 use crate::session::WsSession;
